Trailing slash should catch directories
Found an issue where if you provide a .gitignore line with a trailing slash ('foo/'), it will match files inside but not the directory itself. The .gitignore spec specifies that the directory should be matched as well. This commit changes the end of spec regex to * to fix this issue.